benchmarks/serving_benchmark/requirements.txt
# Dependencies for LLM serving benchmark
# Install: pip install -r benchmarks/serving_benchmark/requirements.txt

numpy>=1.24.0
pandas>=2.0.0
pyarrow>=14.0.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.13.0
tqdm>=4.65.0
tabulate>=0.9.0

# Optional: GPU monitoring (requires NVIDIA drivers)
# pynvml>=11.5.0
